pyproject.toml
[project]
name = "ipdata"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "tenacity (>=9.0.0,<10.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "furl (>=2.1.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
httpx = "^0.28.1"
pytest-cov = "^6.0.0"
pre-commit = "^4.1.0"
autoflake = "^2.3.1"
pytest-mock = "^3.14.0"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120